backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
python-jose
passlib[bcrypt]
python-dotenv
